% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneer.R
\name{create_knee_locator}
\alias{create_knee_locator}
\title{Initialization function for KneeLocator class}
\usage{
create_knee_locator(x, y, S = 1, smooth = "loess")
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{S}{Sensitivity parameter. Smaller values detect knees quicker,
while larger values are more conservative.}

\item{smooth}{Method for smoothing the curve - either smooth.spline or loess}
}
\description{
Initialization function for KneeLocator class
}
\examples{
knee_locator <- create_knee_locator(
			     x = 1:100, y=sapply(1:100, function (x) x**2 - 2*x), smooth="loess")
}
