% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneer.R
\name{find_knee}
\alias{find_knee}
\title{Find Knee.}
\usage{
find_knee(object)
}
\arguments{
\item{object}{KneeLocator object.}
}
\description{
This function runs on the preprocessed KneeLocator object, and finds knees using the
Kneedle algorithm by Satopaa et al. In a nutshell, it works by finding points of maximum
curvature in a curve by finding extrema of the difference curve.
}
\examples{
kl <- create_knee_locator(x=1:10, y=sapply(1:10, function(x) x**3));
			 kl <- preprocess(kl); kl <- find_knee(kl)
}
