% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneer.R
\name{knee_locator_pipeline}
\alias{knee_locator_pipeline}
\title{KneeLocator pipeline}
\usage{
knee_locator_pipeline(x, y, S = 1, smooth = "loess")
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{S}{Sensitivity parameter. Smaller values detect knees quicker,
while larger values are more conservative.}

\item{smooth}{Method for smoothing the curve - either smooth.spline or loess}
}
\description{
This function runs the entire knee locator pipeline, from creation to preprocessing to
finding knees in data. It takes the same inputs as create_knee_locator.
}
\examples{
kl <- knee_locator_pipeline(x=1:10, y=sapply(1:10, function(x) x**3))
}
