% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneer.R
\docType{class}
\name{KneeLocator-class}
\alias{KneeLocator-class}
\alias{KneeLocator}
\title{KneeLocator class}
\description{
The KneeLocator class stores the input data, intermediate calculations and
results of all computations.
}
\section{Slots}{

\describe{
\item{\code{x}}{x coordinates}

\item{\code{y}}{y coordinates}

\item{\code{S}}{Sensitivity parameter. Smaller values detect knees quicker,
while larger values are more conservative.}

\item{\code{smooth}}{Method for smoothing the curve - either smooth.spline or loess}

\item{\code{curve}}{convex or concave}

\item{\code{direction}}{decreasing or increasing}

\item{\code{y.smooth}}{Smoothed values for y coordinates}

\item{\code{x.normalized}}{Min-max scaled x coordinates}

\item{\code{y.normalized}}{Min-max scaled y coordinates}

\item{\code{x.difference}}{x coordinates for difference curve}

\item{\code{y.difference}}{y coordinates for difference curve}

\item{\code{x.difference.maxima}}{maxima locations for difference curve}

\item{\code{y.difference.maxima}}{maxima locations for difference curve}

\item{\code{x.difference.minima}}{minima locations for difference curve}

\item{\code{y.difference.minima}}{momima locations for difference curve}

\item{\code{N}}{number of x, y values}

\item{\code{threshold}}{Threshold for determining maxima}

\item{\code{all_knees}}{List of knees of curve}

\item{\code{all_norm_knees}}{List of knees for normalized curve}

\item{\code{all_knees_y}}{List of y coordinates of knee of curve}

\item{\code{all_norm_knees_y}}{List of y coordinates of knee for normalized curve}

\item{\code{maxima_indices}}{Indices for maxima}

\item{\code{minima_indices}}{Indices for minima}

\item{\code{knee}}{Final knee result - x coordinate}

\item{\code{y_at_knee}}{Final knee result - y coordinate}
}}

