% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneer.R
\docType{class}
\name{KneeLocator}
\alias{KneeLocator}
\alias{KneeLocator,KneeLocator-class}
\title{KneeLocator class}
\description{
The KneeLocator class stores the input data, intermediate calculations and
results of all computations.
}
\section{Slots}{

\describe{
\item{\code{x}}{x coordinates}

\item{\code{y}}{y coordinates}

\item{\code{S}}{Sensitivity parameter. Smaller values detect knees quicker,
while larger values are more conservative.}

\item{\code{smooth}}{Method for smoothing the curve - either smooth.spline or loess}

\item{\code{curve}}{convex or concave}

\item{\code{direction}}{decreasing or increasing}
}}

