% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{maxima}
\alias{maxima}
\title{Calculate maxima of a 1-D vector.}
\usage{
maxima(x)
}
\arguments{
\item{x}{one dimensional vector of floats or ints}
}
\description{
Adapted from https://stackoverflow.com/questions/6836409/finding-local-maxima-and-minima
}
\examples{
maxima(x = c(0, 1, 2, 3, 1, 0))
}
