% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneer.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess}
\usage{
preprocess(object)
}
\arguments{
\item{object}{KneeLocator object.}
}
\description{
The preprocess function calculates intermediate values for coordinates that are used
in downstream calculations of the knee. This includes 1) smoothing the y values; 
2) normalizing the curve to be between 0 and 1; 3) transforms the curve to look like a concave
increasing curve that is easy to do computations on; 4) calculate a difference curve - the
extrema of which will be used to find the knees; 5) finds these extrema; 6) calculates a 
threshold for drop off in values - which is used in determining if a knee has been found.
}
\examples{
knee_locator <- create_knee_locator(x=1:10, y=sapply(1:10, function(x) x**3)); 
			 knee_locator <- preprocess(knee_locator)
}
