% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{smooth_y}
\alias{smooth_y}
\title{Smooth a curve given x and y coordinates.}
\usage{
smooth_y(x, y, method = "loess")
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{method}{Method to use for smoothing the curve. One of "loess" or "smooth.spline"}
}
\description{
Smooth a curve given x and y coordinates.
}
\examples{
smooth_y (x = 1:10, y = sapply(1:10, function(x) x**2) + runif(10, -1, 1) , method = "loess")
}
