% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{StatKneeX}
\alias{StatKneeX}
\title{Stat Knee X}
\format{
An object of class \code{StatKnee} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatKneeX
}
\description{
Class for when the x coordinate of knee needs to be drawn.
}
\keyword{datasets}
