% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{StatKneeY}
\alias{StatKneeY}
\title{Stat Knee Y}
\format{
An object of class \code{StatKnee} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatKneeY
}
\description{
Class for when the y coordinate of knee needs to be drawn.
}
\keyword{datasets}
