% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_xy}
\alias{transform_xy}
\title{Transform x, y}
\usage{
transform_xy(x, y, direction = "increasing", curve = "concave")
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{direction}{direction of the curve}

\item{curve}{curvature - convex or concave}
}
\description{
Transform provided x and y coordinates into a concave increasing curve, to simplify
downstream calculations
}
\examples{
transform_xy(x = 1:10, 
              y = sapply(1:10, function(x) x**2), direction = "increasing", curve = "convex")
}
